drivers: thermal: step_wise: Optimize mitigation evaluation
There is a possibility that a zone will have a polling delay to evaluate
the temperature and determine the mitigation temperature. There is a
possibility that the cooling device to mitigate in this zone might be
mitigated by another zone. The non-mitigating polling type thermal zone,
will use the current mitigation for the cooling device and reduces the
mitigation by one level and this level is active till the other
mitigation clears. Once the other mitigation clears, the mitigation vote
by this non-mitigating thermal zone will reduce step by step. Thus the
mitigation is not cleared right away in this case, instead it takes
multiple polling iteration to clear. This could cause performance
impact.

To avoid this scenario in step-wise algorithm, skip the evaluation of
mitigation when the trip is not reached and there is no previous
mitigation vote. This optimization will prevent the unnecessary
mitigation vote from the non-mitigating thermal zone.

Change-Id: I3939cb53937b650ecf432989f065cfa619297397
Signed-off-by: Ram Chandrasekar <rkumbako@codeaurora.org>
Signed-off-by: celtare21 <celtare21@gmail.com>
Signed-off-by: PrimoDev23 <lexx.ps2711@gmail.com>
Signed-off-by: Panchajanya1999 <panchajanya@azure-dev.live>
Signed-off-by: Mrinal Ghosh <mg712702@gmail.com>
Signed-off-by: Kyvangka1610 <kyvangka2002@gmail.com>
Signed-off-by: ukashah1 <ukashah60@gmail.com>